fix: Add polling and extract usePositions hook
