+++ apps/web/src/features/positions/hooks/usePositions.ts
@@ -1,0 +1,21 @@
+import useChainId from '@/hooks/useChainId'
+import useSafeInfo from '@/hooks/useSafeInfo'
+import { useAppSelector } from '@/store'
+import { selectCurrency } from '@/store/settingsSlice'
+import { usePositionsGetPositionsV1Query } from '@safe-global/store/gateway/AUTO_GENERATED/positions'
+
+const POLLING_INTERVAL = 300_000 // 5 minutes
+
+const usePositions = () => {
+  const chainId = useChainId()
+  const { safeAddress } = useSafeInfo()
+  const currency = useAppSelector(selectCurrency)
+  const { currentData } = usePositionsGetPositionsV1Query(
+    { chainId, safeAddress, fiatCode: currency },
+    { skip: !safeAddress || !chainId || !currency, pollingInterval: POLLING_INTERVAL },
+  )
+
+  return currentData
+}
+
+export default usePositions
