+++ apps/web/src/features/positions/hooks/usePositionsFiatTotal.ts
@@ -1,0 +1,11 @@
+import usePositions from '@/features/positions/hooks/usePositions'
+
+const usePositionsFiatTotal = () => {
+  const protocols = usePositions()
+
+  if (!protocols) return 0
+
+  return protocols.reduce((acc, protocol) => acc + Number(protocol.fiatTotal), 0)
+}
+
+export default usePositionsFiatTotal
